feat: update page to use new ServerCard component

fix: add missing import
