+++ client/dashboard/src/pages/home/Home.tsx
@@ -1,5 +1,4 @@
 import { Page } from "@/components/page-layout";
-import { UpdatedAt } from "@/components/updated-at";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import {
@@ -8,31 +7,15 @@ import {
   SelectItem,
   SelectTrigger,
 } from "@/components/ui/select";
-import { Card } from "@/components/ui/card";
-import { CopyButton } from "@/components/ui/copy-button";
-import { MoreActions } from "@/components/ui/more-actions";
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@/components/ui/popover";
-import { cn } from "@/lib/utils";
 import { useTelemetry } from "@/contexts/Telemetry";
 import { useRoutes } from "@/routes";
 import { ToolsetEntry, Toolset } from "@gram/client/models/components";
-import {
-  useListToolsets,
-  useUpdateToolsetMutation,
-} from "@gram/client/react-query";
-import { useQueryClient } from "@tanstack/react-query";
+import { useListToolsets } from "@gram/client/react-query";
 import { useSdkClient } from "@/contexts/Sdk";
-import { Stack } from "@speakeasy-api/moonshine";
+import { ServerCard } from "@/components/server-card";
 import {
-  CheckCircleIcon,
-  LockIcon,
   BookOpenIcon,
   MessageCircleIcon,
-  ExternalLinkIcon,
   PencilRulerIcon,
   ArrowUpIcon,
   SearchIcon,
@@ -604,293 +587,5 @@ function HomeContent() {
         </Page.Section.Body>
       </Page.Section>
     </>
-  );
-}
-
-function ServerCard({
-  toolset,
-  className,
-}: {
-  toolset: ToolsetEntry | undefined;
-  className?: string;
-}) {
-  const routes = useRoutes();
-  const telemetry = useTelemetry();
-  const { url: mcpUrl, pageUrl } = useMcpUrl(toolset);
-  const updateToolsetMutation = useUpdateToolsetMutation();
-  const queryClient = useQueryClient();
-
-  if (!toolset) return null;
-
-  const handlePrivacyToggle = async (isPublic: boolean) => {
-    try {
-      await updateToolsetMutation.mutateAsync({
-        request: {
-          slug: toolset.slug,
-          updateToolsetRequestBody: {
-            mcpIsPublic: isPublic,
-          },
-        },
-      });
-
-      // Invalidate the toolsets query to refresh the data
-      await queryClient.invalidateQueries({
-        queryKey: ["@gram/client", "toolsets", "list"],
-      });
-
-      telemetry.capture("home_action", {
-        action: "privacy_toggle",
-        server: toolset.slug,
-        newState: isPublic ? "public" : "private",
-      });
-    } catch (error) {
-      // Error is handled by the mutation's error state
-      console.error("Failed to toggle privacy:", error);
-    }
-  };
-
-  return (
-    <Card
-      className={cn(className, "group transition-colors hover:bg-muted/50")}
-    >
-      <Card.Header
-        className="cursor-pointer items-start"
-        onClick={() => routes.toolsets.toolset.goTo(toolset.slug)}
-      >
-        <div className="flex flex-col gap-1">
-          <Card.Title className="text-base">
-            <div className="flex items-center gap-1 group">
-              {toolset.name}
-              <CopyButton
-                text={toolset.slug}
-                size="icon-sm"
-                tooltip="Copy slug"
-                className="text-muted-foreground/80 hover:text-foreground opacity-0 group-hover:opacity-100"
-              />
-            </div>
-          </Card.Title>
-          <UpdatedAt date={new Date(toolset.updatedAt)} italic={false} />
-        </div>
-        <MoreActions
-          actions={[
-            {
-              label: "Manage Tools",
-              onClick: () => routes.toolsets.toolset.goTo(toolset.slug),
-              icon: "blocks",
-            },
-            {
-              label: "MCP Settings",
-              onClick: () => routes.mcp.details.goTo(toolset.slug),
-              icon: "cog",
-            },
-          ]}
-        />
-      </Card.Header>
-      <Card.Content
-        className="cursor-pointer"
-        onClick={() => routes.toolsets.toolset.goTo(toolset.slug)}
-      >
-        <Card.Description>
-          A toolset created from your OpenAPI document
-        </Card.Description>
-      </Card.Content>
-      <Card.Footer>
-        <Stack direction="horizontal" gap={2} align="center">
-          {/* Status Badge with Privacy Controls */}
-          <Popover>
-            <PopoverTrigger asChild>
-              {toolset.mcpIsPublic ? (
-                <Badge
-                  variant="outline"
-                  size="sm"
-                  className="cursor-pointer hover:bg-green-50 transition-colors flex items-center gap-1 uppercase font-mono"
-                  onClick={(e) => e.stopPropagation()}
-                >
-                  <CheckCircleIcon className="w-3 h-3 text-green-500" />
-                  Public
-                </Badge>
-              ) : (
-                <Badge
-                  variant="outline"
-                  size="sm"
-                  className="cursor-pointer hover:bg-muted/50 transition-colors flex items-center gap-1 uppercase font-mono"
-                  onClick={(e) => e.stopPropagation()}
-                >
-                  <LockIcon className="w-3 h-3 text-muted-foreground" />
-                  Private
-                </Badge>
-              )}
-            </PopoverTrigger>
-            <PopoverContent
-              className="w-80"
-              align="start"
-              onClick={(e) => e.stopPropagation()}
-            >
-              <div className="space-y-6">
-                {/* Privacy Toggle Section */}
-                <div className="space-y-3">
-                  <div className="flex items-center justify-between">
-                    <div>
-                      <h4 className="text-base font-light">Server Privacy</h4>
-                      <p className="text-xs text-muted-foreground">
-                        {updateToolsetMutation.isPending
-                          ? "Updating privacy settings..."
-                          : toolset.mcpIsPublic
-                          ? "Publicly accessible and installable"
-                          : "Private, only accessible to you"}
-                      </p>
-                    </div>
-                    <button
-                      onClick={() => handlePrivacyToggle(!toolset.mcpIsPublic)}
-                      disabled={updateToolsetMutation.isPending}
-                      aria-label={`Toggle server privacy. Currently ${
-                        toolset.mcpIsPublic ? "public" : "private"
-                      }`}
-                      role="switch"
-                      aria-checked={toolset.mcpIsPublic}
-                      className={cn(
-                        "relative inline-flex h-5 w-9 items-center rounded-full transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50",
-                        toolset.mcpIsPublic ? "bg-primary" : "bg-input"
-                      )}
-                    >
-                      <span
-                        className={cn(
-                          "pointer-events-none block h-4 w-4 rounded-full bg-background shadow-lg ring-0 transition-transform",
-                          toolset.mcpIsPublic
-                            ? "translate-x-4"
-                            : "translate-x-0.5"
-                        )}
-                      />
-                    </button>
-                  </div>
-                </div>
-
-                <div className="flex flex-col gap-3 pt-3">
-                  {/* MCP URL Section */}
-                  {mcpUrl && (
-                    <div className="space-y-2">
-                      <label className="text-xs text-muted-foreground">
-                        MCP URL
-                      </label>
-                      <div className="flex items-center gap-3">
-                        <code className="flex-1 text-xs bg-muted/50 px-2 py-1 rounded border text-muted-foreground font-mono overflow-x-auto whitespace-nowrap min-w-0">
-                          {mcpUrl}
-                        </code>
-                        <div className="flex-shrink-0">
-                          <CopyButton text={mcpUrl} size="icon-sm" />
-                        </div>
-                      </div>
-                    </div>
-                  )}
-
-                  {/* Install URL Section */}
-                  <div className="space-y-2">
-                    <label className="text-xs text-muted-foreground">
-                      Install Page
-                    </label>
-                    {toolset.mcpIsPublic && pageUrl ? (
-                      <div className="flex items-center gap-3">
-                        <code className="flex-1 text-xs bg-muted/50 px-2 py-1 rounded border text-muted-foreground font-mono overflow-x-auto whitespace-nowrap min-w-0">
-                          {pageUrl}
-                        </code>
-                        <div className="flex-shrink-0">
-                          <CopyButton text={pageUrl} size="icon-sm" />
-                        </div>
-                      </div>
-                    ) : (
-                      <>
-                        {pageUrl ? (
-                          <div className="flex items-center gap-3">
-                            <code className="flex-1 text-xs bg-muted/50 px-2 py-1 rounded border text-muted-foreground font-mono overflow-x-auto whitespace-nowrap min-w-0">
-                              {pageUrl}
-                            </code>
-                            <div className="flex-shrink-0">
-                              <CopyButton text={pageUrl} size="icon-sm" />
-                            </div>
-                          </div>
-                        ) : (
-                          <div className="flex items-center gap-3">
-                            <code className="flex-1 text-xs bg-muted/30 px-2 py-1 rounded border border-dashed text-muted-foreground font-mono">
-                              Will be generated when public
-                            </code>
-                          </div>
-                        )}
-                        <p className="text-xs text-muted-foreground">
-                          Install page will be available once server is made
-                          public
-                        </p>
-                      </>
-                    )}
-                  </div>
-                </div>
-
-                <div className="pt-2">
-                  <Button
-                    variant="outline"
-                    size="sm"
-                    onClick={(e) => {
-                      e.preventDefault();
-                      routes.mcp.details.goTo(toolset.slug);
-                    }}
-                    className="w-full"
-                    caps
-                  >
-                    Advanced Settings
-                  </Button>
-                </div>
-              </div>
-            </PopoverContent>
-          </Popover>
-
-          {/* Quick Test Badge */}
-          <div
-            onClick={(e) => {
-              e.preventDefault();
-              e.stopPropagation();
-              telemetry.capture("home_action", {
-                action: "server_card_action",
-                button: "playground",
-                server: toolset.slug,
-              });
-              routes.playground.goTo(toolset.slug);
-            }}
-          >
-            <Badge
-              variant="outline"
-              size="sm"
-              className="cursor-pointer hover:bg-muted/50 transition-colors flex items-center gap-1 uppercase font-mono"
-            >
-              <MessageCircleIcon className="w-3 h-3" />
-              Test
-            </Badge>
-          </div>
-
-          {/* Install Badge for Public Servers */}
-          {toolset.mcpIsPublic && pageUrl && (
-            <div
-              onClick={(e) => {
-                e.preventDefault();
-                e.stopPropagation();
-                window.open(pageUrl, "_blank");
-                telemetry.capture("home_action", {
-                  action: "server_card_action",
-                  button: "install",
-                  server: toolset.slug,
-                });
-              }}
-            >
-              <Badge
-                variant="outline"
-                size="sm"
-                className="cursor-pointer hover:bg-muted/50 transition-colors flex items-center gap-1 uppercase font-mono"
-              >
-                <ExternalLinkIcon className="w-3 h-3" />
-                Install
-              </Badge>
-            </div>
-          )}
-        </Stack>
-      </Card.Footer>
-    </Card>
   );
 }
